+++ package.json
@@ -63,7 +63,7 @@
     "@canonical/cookie-policy": "3.2.0",
     "@canonical/global-nav": "2.5.0",
     "@canonical/latest-news": "1.2.0",
-    "@canonical/react-components": "^0.17.0",
+    "@canonical/react-components": "^0.17.1",
     "@reduxjs/toolkit": "^1.5.0",
     "date-fns": "2.17.0",
     "react": "^17.0.2",
